+++ np.MeetingNotes/src/NPMeetingNotes.js
@@ -66,6 +66,7 @@ export async function insertNoteTemplate(origFileName: string, dailyNoteDate: Da
         note.content = result
       }
     }
+    logDebug(pluginJson, `MeetingNotes: finished applying rendered template to daily note: ${note?.title || ''}`)
   } else {
     logDebug(pluginJson, `MeetingNotes: got rendered content back; applying to the current editor: shouldReplaceContent:${String(shouldReplaceContent)}`)
     if (shouldReplaceContent) {
@@ -82,6 +83,7 @@ export async function insertNoteTemplate(origFileName: string, dailyNoteDate: Da
     } else {
       Editor.insertTextAtCursor(result)
     }
+    logDebug(pluginJson, `MeetingNotes: finished applying rendered template to Editor (not Editor.note)`)
   }
 }
 
@@ -562,14 +564,14 @@ async function newNoteWithFolder(content: string, _folder?: string): Promise<?st
       }
     }
 
-    logDebug(pluginJson, 'create a new note')
+    logDebug(pluginJson, `creating a new note in folder: "${folder || ''}"`)
     // $FlowFixMe
     const filename = DataStore.newNoteWithContent(content, folder)
 
-    logDebug(pluginJson, 'open the created note')
+    logDebug(pluginJson, `opening the created note: "${filename}"`)
     Editor.openNoteByFilename(filename)
 
-    logDebug(pluginJson, 'find the note and return the title')
+    logDebug(pluginJson, `finding the note and returning the title: "${filename}"`)
     const note = DataStore.projectNoteByFilename(filename)
     if (note) {
       return note.title