+++ np.Templating/CHANGELOG.md
@@ -6,6 +6,17 @@ See Plugin [Documentation](https://noteplan.co/templates/docs) for details on av
 
 DBW: REMEMBER THAT IF YOU ADDED ANY HELPERS IMPORTS, ADD THEM TO THE HELPER MODULE TO GIVE SCRIPTS ACCESS TO THEM ALSO
 
+## [2.1.1] 2025-09-08 @dwertheimer
+- fix using button on empty template, especially with teamspace notes
+- use chooseFolder() to include teamspaces in folder selection of templateNew command
+- add chooseNoteV2() to helpers module
+- fix edge case in date.format() where it was not preserving timezone information in some cases
+- fix other issues with DateModule and TimeModule where it was not preserving timezone information in some cases (using Date objects instead of moment objects -- now using moment objects exclusively)
+- Add smart quotes replacement in import/include files
+- Fix edge case where inline title was not being rendered if it contained EJS tags
+### b3
+- fix chooseFolder() that was forcing new folder creation when it was not needed
+
 ## [2.1.0] 2025-08-?? @dwertheimer
 
 ## New Features